advanced_task 1 edited

+++ 100-clean_web_static.py
@@ -21,12 +21,13 @@ def do_clean(number=0):
         created_at.append(times)
     created_at = sorted(created_at)
     created_at = created_at[len(created_at) - 2:len(created_at)]
-    if number == 0 or number == 1:
+    # print(type(number))
+    if int(number) == 0 or int(number) == 1:
         created_at = [created_at[-1]]
     # print(created_at)
     for i in os.listdir("versions"):
         times = datetime.fromtimestamp(os.stat("versions/" + i).st_ctime)
-        if times not in list(created_at):
+        if times not in created_at:
             local("rm versions/{}".format(i))
             # print(i)
     names = []
@@ -41,12 +42,14 @@ def do_clean(number=0):
              /data/web_static/shared/".format(names[0]))
         sudo("mv /data/web_static/releases/{}\
              /data/web_static/shared/".format(names[1]))
-        sudo("rm -r /data/web_static/releases/*")
+        if len(names) > 1:
+            sudo("rm -r /data/web_static/releases/*")
         # sudo("mkdir /data/web_static/releases/")
         # sudo("chown ubuntu:ubuntu releases")
         sudo("mv /data/web_static/shared/* /data/web_static/releases")
     except IndexError:
-        sudo("rm -r /data/web_static/releases/*")
+        if len(names) > 1:
+             sudo("rm -r /data/web_static/releases/*")
         sudo("mv /data/web_static/shared/* /data/web_static/releases")
     # print(list_files)
     # print(names[0])
